Jacob and API cleanup
Cleans up the api folder and import logic

Jacob now controls the api function since it's used within that function

Increased log information, might be spammy but will allow for console only debugging.

Changed the way preload is loaded, I intend to change that for other features as well (don't want to touch main stuff due to not being able to test tools etc.)
