+++ .gitignore
@@ -13,3 +13,5 @@
 
 # Personal Config File
 config.toml
+
+jsconfig.json
